+++ app/Http/Controllers/GradeController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\FireFighter;
 use App\Models\Grade;
 use App\Models\State;
 use Illuminate\Http\Request;
@@ -31,13 +32,30 @@ class GradeController extends Controller
     public function delete($id)
     {
         $grade = Grade::findOrFail($id);
-        $grade->delete();
+
+        $fireFighters = FireFighter::all();
+
+        $isUsed = false;
+
+        foreach ($fireFighters as $ff)
+        {
+            if ($ff->idGrade == $grade->id)
+                $isUsed = true;
+        }
+
+        if (!$isUsed)
+            $grade->delete();
+        
         return redirect()->back();
     }
 
     public function clear()
     {
-        Grade::query()->delete();
+        $fireFighters = FireFighter::all();
+
+        if ($fireFighters->count() == 0)
+            Grade::query()->delete();
+        
         return redirect()->back();
     }
 }
